% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpprbinom}
\alias{cpprbinom}
\title{return a binomial distributed vector}
\usage{
cpprbinom(n, size, prob)
}
\arguments{
\item{n}{number of rows of X}

\item{size}{1}

\item{prob}{means of output vector}
}
\description{
return a binomial distributed vector
}
