% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{genDataList}
\alias{genDataList}
\title{Generated X from a p-dimensional normal distribution with mean 0p and covariance matrix
and y = Xβ + ε
β is sampled i.i.d. from N (0, 1) and then projected onto the row space of X to ensure identifiability}
\usage{
genDataList(n, mu, p, rho, beta, SNR, family)
}
\arguments{
\item{n}{number of rows of X}

\item{mu}{a zero vector of length p where p is number of cols of X}

\item{p}{number of cols of X}

\item{rho}{where Σuv = rho to the power of |u−v|}

\item{beta}{β is sampled i.i.d. from N (0, 1) with length p}

\item{SNR}{signal-to-noise ratio for, 10 for gaussian family case and NaN for poisson and binomial cases}

\item{family}{three family cases: gaussian, poisson and binomial}
}
\description{
Generated X from a p-dimensional normal distribution with mean 0p and covariance matrix
and y = Xβ + ε
β is sampled i.i.d. from N (0, 1) and then projected onto the row space of X to ensure identifiability
}
