% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GradientLs}
\alias{GradientLs}
\title{GradientLs function}
\usage{
GradientLs(theta, X, y, family)
}
\arguments{
\item{theta}{unknown regression vector}

\item{X}{high dimensional matrix}

\item{y}{where y = Xβ + ε}

\item{family}{three family cases: gaussian, poisson and binomial}
}
\description{
GradientLs function
}
