% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lqnorm.R
\name{Lqnorm}
\alias{Lqnorm}
\title{Lqnorm function}
\usage{
Lqnorm(q, v)
}
\arguments{
\item{q}{integer}

\item{v}{vector}
}
\value{
a real number which is q's norm of a vector
}
\description{
generate q's norm of a vector
}
\details{
q-norm (also called lq-norm)
}
