% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MeanFunction}
\alias{MeanFunction}
\title{mean function}
\usage{
MeanFunction(X, theta, family)
}
\arguments{
\item{X}{high dimensional matrix}

\item{theta}{unknown regression vector}

\item{family}{three family cases: gaussian, poisson and binomial}
}
\description{
mean function
}
