% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MeanPrime}
\alias{MeanPrime}
\title{mean prime function}
\usage{
MeanPrime(X, theta, family)
}
\arguments{
\item{X}{high dimensional matrix}

\item{theta}{unknown regression vector}

\item{family}{three family cases: gaussian, poisson and binomial}
}
\description{
mean prime function
}
