% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvrnormArma}
\alias{mvrnormArma}
\title{Generate a design matrix and normalize each column of the design matrix}
\usage{
mvrnormArma(n, mu, p, rho)
}
\arguments{
\item{n}{number of rows of X}

\item{mu}{a zero vector of length p where p is number of cols of X}

\item{p}{number of columns of X}

\item{rho}{where Σuv = rho to the power of |u−v|}
}
\description{
Generate a design matrix and normalize each column of the design matrix
}
