% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optim_ObFn}
\alias{optim_ObFn}
\title{find optimized solution for Objective function}
\usage{
optim_ObFn(theta, X, y, family, trex_c)
}
\arguments{
\item{theta}{vector store our optimized result}

\item{X}{high dimensional matrix}

\item{y}{where y = Xβ + ε}

\item{family}{three family cases: gaussian, poisson and binomial}

\item{trex_c}{TREX parameters, 2 for gaussian case, 1 for poisson and binomial case}
}
\description{
find optimized solution for Objective function
}
