% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optim_ObLs}
\alias{optim_ObLs}
\title{find optimized solution for ObjectLs function}
\usage{
optim_ObLs(theta, X, y, family)
}
\arguments{
\item{theta}{vector store our optimized result}

\item{X}{high dimensional matrix}

\item{y}{where y = Xβ + ε}

\item{family}{three family cases: gaussian, poisson and binomial}
}
\description{
find optimized solution for ObjectLs function
}
