% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optim_Ridge}
\alias{optim_Ridge}
\title{find optimized solution for ObjectRidge function}
\usage{
optim_Ridge(theta, X, y, family, r)
}
\arguments{
\item{theta}{vector store our optimized result}

\item{X}{high dimensional matrix}

\item{y}{where y = Xβ + ε}

\item{family}{three family cases: gaussian, poisson and binomial}

\item{r}{tuning.parameter}
}
\description{
find optimized solution for ObjectRidge function
}
