% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rpois_rcpp}
\alias{rpois_rcpp}
\title{return a poisson distributed vector}
\usage{
rpois_rcpp(lambda)
}
\arguments{
\item{lambda}{a vector which is the mean of result vector}
}
\description{
return a poisson distributed vector
}
