% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Test.case_gau}
\alias{Test.case_gau}
\title{A Test case
selected from three different distributions}
\format{
\describe{get by calling TridgeEst function
\item{Test.case_gau}{input test case, here "gaussian"}
...
}
}
\usage{
Test.case_gau
}
\description{
A Test case
selected from three different distributions
}
\keyword{datasets}
