% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Test.case_poi}
\alias{Test.case_poi}
\title{A Test case
selected from three different distributions}
\format{
\describe{get by calling TridgeEst function
\item{Test.case_poi}{input test case, here "poisson"}
...
}
}
\usage{
Test.case_poi
}
\description{
A Test case
selected from three different distributions
}
\keyword{datasets}
