% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TridgeEst.R
\name{TridgeEst}
\alias{TridgeEst}
\title{Tridge Estimator}
\usage{
TridgeEst(Test.case, num.obs, num.par, k)
}
\arguments{
\item{Test.case}{three different function distributions}

\item{num.obs}{num of rows of the matrix X}

\item{num.par}{num of volomns of the matrix y}

\item{k}{magnitude of the mutual correlations}
}
\value{
tridge estimator vector
}
\description{
generate a tridge estimator for objective function
}
\details{
DETAILS
}
\examples{
TridgeEst(Test.case="gaussian",num.obs=100,num.par=300,k=0.0)
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
