% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{y_bin}
\alias{y_bin}
\title{y vector,with ,dimension=X.rows
A dataset randomly generate the output vector}
\format{
\describe{get by calling TridgeEst function
\item{y_bin}{input vector,dimension=10}
...
}
}
\usage{
y_bin
}
\description{
y vector,with ,dimension=X.rows
A dataset randomly generate the output vector
}
\keyword{datasets}
