% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{y_poi}
\alias{y_poi}
\title{y vector,with ,dimension=X.rows
A dataset randomly generate the output vector}
\format{
\describe{get by calling TridgeEst function
\item{y_poi}{input vector,dimension=100}
...
}
}
\usage{
y_poi
}
\description{
y vector,with ,dimension=X.rows
A dataset randomly generate the output vector
}
\keyword{datasets}
