% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativeError.R
\name{relativeError}
\alias{relativeError}
\title{relative error table}
\usage{
relativeError(Test.case, num.obs, num.par, k, num.runs)
}
\arguments{
\item{Test.case}{three different function distributions}

\item{num.obs}{num of rows of the matrix X}

\item{num.par}{num of volomns of the matrix y}

\item{k}{magnitude of the mutual correlations}

\item{num.runs}{num of runs to estimate and compute error}
}
\value{
a table showing the errors for different methods
}
\description{
compare errors of estimator for tridge with MLE or 10 fold CV
}
\details{
call TridgeEst to computer the t-ridge estimator then compare the relative errors
}
\examples{
Test.case="gaussian"
num.obs=100
num.par=300
k=0.0
num.runs=5
}
