% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TridgeEst.R
\name{TridgeEst}
\alias{TridgeEst}
\title{Tridge Estimator}
\usage{
TridgeEst(Test.case, X, y)
}
\arguments{
\item{Test.case}{three different function distributions}

\item{X}{matrix with num.obs x num.par dimension}

\item{y}{vector, num.obs elements}
}
\value{
tridge estimator vector
}
\description{
generate a tridge estimator for objective function
}
\details{
First compute the stationary point for beta by calling optim_ObFn, then compute the ridge estimaor for m equally spaced tuning
parameters and find the one minimize the objective function.
}
\examples{
Test.case="gaussian"
X=matrix(c(1:12),3,4)
y=c(2,4,5)
}
