% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X_poi}
\alias{X_poi}
\title{matrix X sampled from a 300-dimensional normal distribution under poisson case
A dataset randomly generate the high dimensioanl matrix X}
\format{
\describe{get by calling TridgeEst function
\item{X_poi}{input matrix,(100x300)}
...
}
}
\usage{
X_poi
}
\description{
matrix X sampled from a 300-dimensional normal distribution under poisson case
A dataset randomly generate the high dimensioanl matrix X
}
\keyword{datasets}
